+++ packages/lab/src/input-next/InputNext.tsx
@@ -236,7 +236,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
         onChange={handleChange}
         onKeyDown={onKeyDown}
         onKeyUp={onKeyUp}
-        onFocus={!isDisabled && handleFocus}
+        onFocus={!isDisabled ? handleFocus : undefined}
         onMouseDown={handleMouseDown}
         onMouseUp={handleMouseUp}
         onMouseMove={handleMouseMove}
